requirements.txt
streamlit
Pillow
rembg[gpu]
#rembg[cpu]
